about.html re-done and looks good. will expand on.
all room images on index.html bring to new pages for each different room. they are all the POOL-SIDE code, so will need to be changed (headers and pragraphs, and images.)

and i did this because finally the room page looks really good colour scheme wise. the text needed to be changed to white FFFFFF instead og grey FFF